# shared/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WedPlan wedding RSVP and asoebi commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkout/flutterwave:
    post:
      operationId: startFlutterwaveCheckout
      tags: [checkout]
      summary: Create a Flutterwave checkout link for an RSVP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Flutterwave checkout link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Invalid checkout request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Flutterwave is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout/bank-transfer:
    post:
      operationId: createBankTransferOrder
      tags: [checkout]
      summary: Create a bank transfer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Bank transfer instructions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransferResponse"
        "400":
          description: Invalid checkout request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Bank transfer is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout/gift/flutterwave:
    post:
      operationId: startGiftFlutterwaveCheckout
      tags: [checkout]
      summary: Create a Flutterwave checkout link for a monetary gift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCheckoutInput"
      responses:
        "200":
          description: Flutterwave checkout link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Invalid gift request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Flutterwave is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkout/gift/bank-transfer:
    post:
      operationId: createGiftBankTransferOrder
      tags: [checkout]
      summary: Create a bank transfer order for a monetary gift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCheckoutInput"
      responses:
        "201":
          description: Bank transfer instructions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransferResponse"
        "400":
          description: Invalid gift request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Bank transfer is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GiftCheckoutInput:
      type: object
      required: [guestName, email, amount]
      properties:
        guestName: { type: string, minLength: 2 }
        email: { type: string }
        amount: { type: number, minimum: 1 }
        message: { type: string }
    CheckoutItemInput:
      type: object
      required: [guestName, itemId, size]
      properties:
        guestName: { type: string }
        itemId: { type: number }
        size: { type: string }
        quantity: { type: number, minimum: 1 }
    CheckoutInput:
      type: object
      required: [rsvpId, guestName, email, items]
      properties:
        rsvpId: { type: number }
        guestName: { type: string }
        email: { type: string }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CheckoutItemInput"
    CheckoutResponse:
      type: object
      required: [reference, checkoutUrl]
      properties:
        reference: { type: string }
        checkoutUrl: { type: string }
    BankTransferResponse:
      type: object
      required: [reference, amount, currency, bankName, accountName, accountNumber, instructions]
      properties:
        reference: { type: string }
        amount: { type: number }
        currency: { type: string }
        bankName: { type: string }
        accountName: { type: string }
        accountNumber: { type: string }
        instructions: { type: string }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...