# shared/src/generated/types/bankTransferResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * WedPlan wedding RSVP and asoebi commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface BankTransferResponse {
  reference: string;
  amount: number;
  currency: string;
  bankName: string;
  accountName: string;
  accountNumber: string;
  instructions: string;
}